observability/grafana/httpmetrics/requirements.txt
statsd==4.0.1
python-dotenv==1.0.0
flask==2.3.2
